Initial WIP Upload
This was just a small project to practise React. It consists of a couple of simple SpringBoot APIs and a React app that calls them to display information on a given vehicle. It does this using the DVLA's public API. I might expand it in the future, I might not.